+++ EFO/src/main/java/com/efo/interfaces/IPaymentsReceived.java
@@ -4,8 +4,8 @@ import com.efo.entity.PaymentsReceived;
 
 public interface IPaymentsReceived {
 	public void create(PaymentsReceived payments);
-	public PaymentsReceived retreive(int id);
+	public PaymentsReceived retreive(Long id);
 	public void update(PaymentsReceived payments);
-	public void delete(int id);
+	public void delete(Long id);
 
 }
